requirements.txt
fastapi
uvicorn
python-dotenv
httpx
Jinja2
python-multipart
